counsel.el (counsel-compile): fall-back to default-directory (#2652)
Some file we compile may not have a project root associated with
them. In this case we should fallback to the default-directory which
for any given file should be it's home directory.

This does confuse counsel-compile-get-filtered-history somewhat when
building something in $HOME as generally all projects will be children
of $HOME.