counsel.el (counsel-compile): fall-back to default-directory (#2652)
Some file we compile may not have a project root associated with
them. In this case we should fallback to the buffers home
directory (which should also be default-directory).

To prevent confusing counsel-compile-get-filtered-history matching a
whole bunch of project history we modify a buffer local variable which
makes counsel-compile-get-filtered-history do a stricter matching
based purely on a matching source directory. This way we get to keep
history for these files.