counsel.el (counsel-compile): fall-back to default-directory (#2652)
Some file we compile may not have a project root associated with
them. In this case we should fallback to the default-directory which
for any given file should be it's home directory.

To prevent confusing counsel-compile-get-filtered-history matching a
whole bunch of project history we do a setq-local for
counsel-compile-history in the compile-root helper.